script: - Increased all stack sizes to 128MB - Fixed bug within chain_str where it was incorrectly comuting size - Introduced flags hints in procedures when bitfields should be concatenated and when they should not
common:
- introduced ALIGN_TO as an alternative to ROUND_UP when using power of 2. It yields better code gen on all compilers.

util:
- Fixed incorrect assignment of HO[0-9] atoms, which were given Ho, now they are identified as a heuristic case and properly mapped to H

bitfield:
- Added reserve_range functiion to reserve a complete range, so you don't have to poke the allocator as much, if you know the size beforehand.

unittest:
- Added some tests for allocators.

+++ src/core/md_file.c
@@ -155,7 +155,7 @@ int64_t md_file_read(md_file_o* file, void* ptr, int64_t num_bytes) {
 
 int64_t md_file_read_line(md_file_o* file, char* buf, int64_t cap) {
     int64_t pos = md_file_tell(file);
-    char* res = fgets(buf, cap, (FILE*)file);
+    char* res = fgets(buf, (int)cap, (FILE*)file);
     int64_t len = md_file_tell(file) - pos;
     return res ? len : 0;
 }
